+++ CHANGES.rst
@@ -11,6 +11,8 @@ Bugfixes
 ^^^^^^^^
 
 - Use correct speaker name when exporting contributions to spreadsheets
+- Use friendly IDs in abstract attachment package folder names
+- Fix typo in material package subcontribution folder names
 
 Version 2.1.4
 -------------